refactor(protocol-jackson): extract common base ValueType ID resolve
Extracts the common behavior of a `TypeIdResolver` which reuses the
value type to resolve a particular type, whether it be an `Intent` type
or a `RecordValue` type.
